pyproject.toml
[project]
name = "texjam"
version = "0.1.0"
description = "A scaffold for LaTeX templates."
authors = [
    {name = "Quaerent",email = "topfyf@qq.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "click (>=8.3.1,<9.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "binaryornot (>=0.4.4,<0.5.0)",
    "appdirs (>=1.4.4,<2.0.0)",
]

[tool.poetry.scripts]
texjam = "texjam.main:cli"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)"
]
